refactor: `share` 모듈 분리
