+++ core/build.gradle
@@ -1,4 +1,6 @@
 dependencies {
+    implementation project(':share')
+
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-jdbc'
