+++ core/src/main/java/com/nilgil/book/core/query/InvalidIsbnException.java
@@ -1,4 +1,4 @@
-package com.nilgil.book.core.share;
+package com.nilgil.book.core.query;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
