+++ core/src/main/java/com/nilgil/book/core/query/Isbn.java
@@ -1,4 +1,4 @@
-package com.nilgil.book.core.share;
+package com.nilgil.book.core.query;
 
 import java.util.Optional;
