+++ core/src/main/java/com/nilgil/book/core/search/BookSearchFacade.java
@@ -8,7 +8,7 @@ import com.nilgil.book.core.search.parser.QueryParser;
 import com.nilgil.book.core.search.parser.model.Query;
 import com.nilgil.book.core.search.planner.PlannedQuery;
 import com.nilgil.book.core.search.planner.QueryPlanner;
-import com.nilgil.book.core.share.PageReq;
+import com.nilgil.book.share.PageReq;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
