+++ core/src/main/java/com/nilgil/book/core/search/executor/MetadataQueryExecutorDecorator.java
@@ -3,7 +3,7 @@ package com.nilgil.book.core.search.executor;
 import com.nilgil.book.core.search.executor.model.BookSearchResult;
 import com.nilgil.book.core.search.executor.model.Metadata;
 import com.nilgil.book.core.search.planner.PlannedQuery;
-import com.nilgil.book.core.share.PageReq;
+import com.nilgil.book.share.PageReq;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
