+++ core/src/main/java/com/nilgil/book/core/search/executor/PostgresFtsQueryExecutor.java
@@ -3,8 +3,8 @@ package com.nilgil.book.core.search.executor;
 import com.nilgil.book.core.search.executor.model.BookHit;
 import com.nilgil.book.core.search.executor.model.BookSearchResult;
 import com.nilgil.book.core.search.planner.PlannedQuery;
-import com.nilgil.book.core.share.PageInfo;
-import com.nilgil.book.core.share.PageReq;
+import com.nilgil.book.share.PageInfo;
+import com.nilgil.book.share.PageReq;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
