--- core/src/main/java/com/nilgil/book/core/share/PageReq.java
@@ -1,12 +1,0 @@
-package com.nilgil.book.core.share;
-
-public record PageReq(
-        int page,
-        int size
-) {
-    public PageReq {
-        page = Math.max(page, 0);
-        size = Math.max(size, 1);
-        size = Math.min(size, 100);
-    }
-}
