+++ core/src/test/java/com/nilgil/book/core/query/BookQueryControllerTest.java
@@ -5,8 +5,7 @@ import com.nilgil.book.core.search.executor.model.BookHit;
 import com.nilgil.book.core.search.executor.model.BookSearchResult;
 import com.nilgil.book.core.search.executor.model.Metadata;
 import com.nilgil.book.core.search.planner.SearchStrategy;
-import com.nilgil.book.core.share.Isbn;
-import com.nilgil.book.core.share.PageInfo;
+import com.nilgil.book.share.PageInfo;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
