+++ settings.gradle
@@ -1,3 +1,4 @@
 rootProject.name = 'book-search'
 
 include 'core'
+include 'share'
