+++ share/src/main/java/com/nilgil/book/share/PageReq.java
@@ -1,0 +1,7 @@
+package com.nilgil.book.share;
+
+public record PageReq(
+        int page,
+        int size
+) {
+}
